TAM 12.3 Funcionalidades de permisos implementadas en los controladores de Accionables, AccionDeMejora, Objetivos, PlanDeMejora y TipoObjetivos.
Se agrega metodo Init para simular el usuario loggeado.

Todos los test que pasaban antes de la implementacion siguen pasando.

+++ AppIntegrador/Controllers/AccionDeMejoraController.cs
@@ -7,16 +7,23 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using AppIntegrador.Models;
+using AppIntegrador.Utilities;
 
 namespace AppIntegrador.Controllers
 {
     public class AccionDeMejoraController : Controller
     {
         private DataIntegradorEntities db = new DataIntegradorEntities();
+        private readonly IPerm permissionManager = new PermissionManager();
 
         // GET: AccionDeMejora
         public ActionResult Index()
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             var objetivo = db.AccionDeMejora.Include(o => o.PlantillaAccionDeMejora).Include(o => o.Objetivo);
             return View("Index", objetivo.ToList());
         }
@@ -24,12 +31,22 @@ namespace AppIntegrador.Controllers
         // GET: AccionDeMejora/Details/5
         public ActionResult Details(int id)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             return View();
         }
 
         // GET: AccionDeMejora/Create
         public ActionResult Create()
         {
+            if (!permissionManager.IsAuthorized(Permission.CREAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             ViewBag.codPlan = new SelectList(db.PlanDeMejora, "codigo", "nombre");
             ViewBag.nombreObj = new SelectList(db.Objetivo, "nombre", "nombre");
             return View("_createAccionDeMejora");
@@ -40,6 +57,11 @@ namespace AppIntegrador.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "codPlan,nombreObj,descripcion,fechaInicio,fechaFin,codPlantilla")] AccionDeMejora accionDeMejora)
         {
+            if (!permissionManager.IsAuthorized(Permission.CREAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             bool error = false;
 
             if (accionDeMejora.fechaInicio != null && accionDeMejora.fechaFin != null)
@@ -94,6 +116,11 @@ namespace AppIntegrador.Controllers
         // Corresponde a MOS 1.3 (2)
         public ActionResult Edit(int plan, string obj, string descripcion)
         {
+            if (!permissionManager.IsAuthorized(Permission.EDITAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             if (plan == null || obj == null || descripcion == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -112,6 +139,11 @@ namespace AppIntegrador.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "codPlan,nombreObj,descripcion,fechaInicio,fechaFin,,codPlantilla")] AccionDeMejora accionDeMejora)
         {
+            if (!permissionManager.IsAuthorized(Permission.EDITAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             bool error = false;
 
             if (accionDeMejora.fechaInicio != null && accionDeMejora.fechaFin != null)
@@ -137,6 +169,11 @@ namespace AppIntegrador.Controllers
         // Corresponde a MOS 1.3 (2)
         public ActionResult Delete(int? plan, string nombObj, string descripcion)
         {
+            if (!permissionManager.IsAuthorized(Permission.BORRAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             if (plan == null || nombObj == null || descripcion == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -155,6 +192,11 @@ namespace AppIntegrador.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int? plan, string nombObj, string descripcion)
         {
+            if (!permissionManager.IsAuthorized(Permission.BORRAR_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             AccionDeMejora accion = db.AccionDeMejora.Find(plan, nombObj, descripcion);
             db.AccionDeMejora.Remove(accion);
             db.SaveChanges();
@@ -185,6 +227,11 @@ namespace AppIntegrador.Controllers
 
         public ActionResult AccionesDeObjetivo(int plan, string nombObj, bool edit = true)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_ACCIONES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             ViewBag.idPlan = plan;
             ViewBag.nombreObj = nombObj;
