+++ AppIntegrador/Controllers/PlanDeMejoraController.cs
@@ -10,27 +10,36 @@ using System.Web;
 using System.Web.Mvc;
 using System.Web.UI;
 using AppIntegrador.Models;
+using AppIntegrador.Utilities;
 
 namespace AppIntegrador.Controllers
 {
     public class PlanDeMejoraController : Controller
     {
         private DataIntegradorEntities db = new DataIntegradorEntities();
+        private readonly IPerm permissionManager;
 
         public PlanDeMejoraController()
         {
             this.db = new DataIntegradorEntities();
+            permissionManager = new PermissionManager();
         }
 
         public PlanDeMejoraController(DataIntegradorEntities mdb)
         {
             this.db = mdb;
+            permissionManager = new PermissionManager();
         }
 
         // GET: PlanDeMejora
         [HttpGet]
         public ActionResult Index(List<PlanDeMejora> planes = null)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             if (planes == null || planes.Count == 0)
             {
                 planes = db.PlanDeMejora.ToList();
@@ -46,6 +55,11 @@ namespace AppIntegrador.Controllers
 
         public ActionResult Buscar(String nombrePlan)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             var planes = db.PlanDeMejora.Where(x => x.nombre.Contains(nombrePlan)).ToList();
             return Index(planes);
         }
@@ -55,6 +69,11 @@ namespace AppIntegrador.Controllers
         */
         public ActionResult Index(String nombre)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             ObjectParameter count = new ObjectParameter("count", 999);
             ViewBag.cantidad = count.Value;
             ViewBag.nombre = nombre;
@@ -69,6 +88,11 @@ namespace AppIntegrador.Controllers
         */
         public ActionResult Crear(PlanDeMejora plan = null)
         {
+            if (!permissionManager.IsAuthorized(Permission.CREAR_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             if (plan == null)
             {
                 plan = new PlanDeMejora();
@@ -94,6 +118,11 @@ namespace AppIntegrador.Controllers
                                     List<String> FormularioSeleccionado = null,
                                     List<Objetivo> MisObjetivos = null)
         {
+            if (!permissionManager.IsAuthorized(Permission.CREAR_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             PlanDeMejora planAgregado = null;
             Profesor profe;
             Formulario formulario;
@@ -133,6 +162,11 @@ namespace AppIntegrador.Controllers
         [HttpPost]
         public ActionResult AnadirProfes(List<String> ProfeSeleccionado)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             List<Profesor> profesores = new List<Profesor>();
             List<String> ProfesoresNombreLista = new List<String>();
             ObjectParameter name_op;
@@ -156,6 +190,11 @@ namespace AppIntegrador.Controllers
         [HttpPost]
         public ActionResult AnadirFormularios(List<String> FormularioSeleccionado)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             List<Formulario> formularios = new List<Formulario>();
             if (FormularioSeleccionado != null)
             {
@@ -174,6 +213,11 @@ namespace AppIntegrador.Controllers
         //retorna la vista de editar para que puedan ser añadidos los objetivos, acciones y acionables al mismo
         public ActionResult EditarPlanDeMejora(int id)
         {
+            if (!permissionManager.IsAuthorized(Permission.EDITAR_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             ViewBag.IdPlan = id;
             PlanDeMejora planDeMejora = db.PlanDeMejora.Find(id);
             ViewBag.Editar = true;
@@ -188,6 +232,11 @@ namespace AppIntegrador.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult EditarPlanDeMejora2([Bind(Include = "codigo,nombre,fechaInicio,fechaFin")] PlanDeMejora planDeMejora)
         {
+            if (!permissionManager.IsAuthorized(Permission.EDITAR_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(planDeMejora).State = EntityState.Modified;
@@ -233,6 +282,11 @@ namespace AppIntegrador.Controllers
         // Method that deletes one "PlanDeMejora"
         public ActionResult BorrarPlan(int codigoPlan)
         {
+            if (!permissionManager.IsAuthorized(Permission.BORRAR_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             PlanDeMejora planDeMejora = db.PlanDeMejora.Find(codigoPlan);
             db.PlanDeMejora.Remove(planDeMejora);
             db.SaveChanges();
@@ -291,6 +345,11 @@ namespace AppIntegrador.Controllers
         //Permite actualizar la vista parcial de objetivos al crear uno nuevo
         public ActionResult RefrescarObjetivos(int Id)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_OBJETIVOS))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página";
+                return RedirectToAction("../Home/Index");
+            }
             ViewBag.IdPlan = Id;
             IEnumerable<AppIntegrador.Models.Objetivo> objetivosDePlan = db.Objetivo.Where(o => o.codPlan == Id);
             return PartialView("_objetivosDelPlan", objetivosDePlan);
@@ -302,6 +361,11 @@ namespace AppIntegrador.Controllers
         [HttpGet]
         public ActionResult Detalles(int id)
         {
+            if (!permissionManager.IsAuthorized(Permission.VER_PLANES_MEJORA))
+            {
+                TempData["alertmessage"] = "No tiene permisos para acceder a esta página.";
+                return RedirectToAction("../Home/Index");
+            }
             ViewBag.IdPlan = id;
             PlanDeMejora planDeMejora = db.PlanDeMejora.Find(id);
             return View("DetallesPlanDeMejora", planDeMejora);
